refactor(tools): widen row counts to long to avoid 2.1B-row overflow

+++ backend/tools/PrmDashboard.CsvExporter/Export/TableExportResult.cs
@@ -7,6 +7,6 @@ namespace PrmDashboard.CsvExporter.Export;
 public sealed record TableExportResult(
     string Label,
     string OutputPath,
-    int RowsWritten,
-    int SourceCount,
+    long RowsWritten,
+    long SourceCount,
     bool Matches);
